% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.conditional_moment}
\alias{augment.conditional_moment}
\title{Augment data with information from a conditional mean fit or conditional variance fit}
\usage{
\method{augment}{conditional_moment}(x, level = 0.95, ...)
}
\arguments{
\item{x}{Model object of class "conditional_moment" returned from
\code{\link[conduits]{conditional_mean}} or  \code{\link[conduits]{conditional_var}}
with information to append to observations.}

\item{level}{Confidence level. Default is set to 0.95.}

\item{...}{Addition arguments to augment method.}
}
\value{
A \code{\link[tibble]{tibble}} with information
about data points.
}
\description{
This function produces partial residuals for each predictor,
and the estimated conditional means, standard error and confidence limits.
}
\examples{
data <- NEON_PRIN_5min_cleaned \%>\%
  dplyr::filter(site == "upstream") \%>\%
  dplyr::select(Timestamp, turbidity, level, conductance, temperature)

fit_mean <- data \%>\%
  conditional_mean(turbidity ~ s(level, k = 8) +
    s(conductance, k = 8) + s(temperature, k = 8))

data_inf <- fit_mean \%>\% augment()
}
\seealso{
\code{\link[mgcv]{gam}}
}
