% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment}
\alias{augment}
\title{Augment data with information from an conditional mean fit}
\usage{
augment(x)
}
\arguments{
\item{x}{Model object returned from conditional_mean
with information to append to observations.}

\item{...}{Addition arguments to augment method.}
}
\value{
A \code{\link[tibble]{tibble}} with information
about data points.
}
\description{
This function produces partial residuals for each predictor,
and the estimated conditional means, standard error and confidence limits.
}
\examples{
mean_fit <- NEON_PRIN_5min_cleaned \%>\%
  dplyr::filter(site == "upstream") \%>\%
  dplyr::select(turbidity, level, conductance, temperature) \%>\%
  conditional_mean(turbidity ~ level + conductance + temperature,
    knots_mean = c(8, 8, 8)
  )

data_inf <- mean_fit \%>\% augment()
}
\seealso{
\code{\link[mgcv]{gam}}
}
