% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.conditional_ccf}
\alias{autoplot.conditional_ccf}
\title{Visualising the models fitted for conditional cross-correlations}
\usage{
\method{autoplot}{conditional_ccf}(object, type = c("mean", "variance", "cross-correlation"), ...)
}
\arguments{
\item{object}{an object returned from \code{conditional_ccf} function}

\item{type}{the type of moment or cross-correlation to visualise.
This can take one of "mean", "variance" or "cross-correlation"}

\item{...}{further arguments passed to or from other methods}
}
\value{
returns plots visualising fitted gam models for conditional
mean or variance vs each predictor.
}
\description{
A function for visualising the fitted generalised additive models for
conditional means, variances and cross-correlations from
\code{conditional_ccf} function
}
\examples{

}
\seealso{
\code{\link[stats]{plot.glm}}, \code{\link[visreg]{visreg}}
}
\author{
Puwasala Gamakumara
}
