% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_ccf.R
\name{autoplot.estimate_dt}
\alias{autoplot.estimate_dt}
\title{Visualising the lag time \eqn{d_t}}
\usage{
\method{autoplot}{estimate_dt}(object, interval = FALSE, m = 100, seed = NULL, ...)
}
\arguments{
\item{object}{an object returned from \code{estimate_dt} function}

\item{interval}{if TRUE, the bootstrap prediction intervals are computed.
This will increase the computation time.Default is setted to FALSE.}

\item{m}{number of replications when getting boostrap prediction intervals
for \eqn{d_t}}

\item{seed}{setting seed}

\item{...}{further arguments passed to or from other methods}
}
\value{
returns plots visualising \eqn{d_t} with each predictor used
in conditional cross-correlation functions. Further it returns a list of
following components
\item{plot_list}{list of plots - ggplot objects}
\item{data}{data used to create individual plots}
}
\description{
A function to visualise the estimated lag time, \eqn{d_t}, with each
of the predictor used is conditional cross-correlation models. The
prediction intervals are computed using the Sieve bootstrap method (Buhlmann 1997)
}
\references{
Bühlmann, P. (1997). Sieve bootstrap for time series. Bernoulli, 3(2), 123-148.
}
\author{
Puwasala Gamakumara
}
