% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dt.R
\name{calc_dt_CI}
\alias{calc_dt_CI}
\title{Computing bootstrapped confidence intervals for dt}
\usage{
calc_dt_CI(x, m, new_data = NULL)
}
\arguments{
\item{x}{Model object of class "conditional_ccf" returned from
\code{\link[conduits]{conditional_ccf}}}

\item{m}{number of replications for boostrap confidence intervals}

\item{new_data}{the dataset with the some predictors that are
set to the median value (if required). Default is set to NULL.}
}
\value{
A \code{\link[tibble]{tibble}} with estimated time lag "dt"
}
\description{
This function computes the  bootstrapped confidence intervals for dt.
It resample the residuals from the various models used in the
conditional cross-correlation calculation to generate new data.
As the  residuals are serially correlated,  a sieve bootstrap
approach to capture the autocorrelation  structure in the data.
}
\examples{
\dontrun{

df_dt <- fit_c_ccf \%>\% calc_dt_CI(2 )

new_data <- fit_c_ccf$data
new_data <- new_data \%>\%
mutate(temperature_upstream = median(temperature_upstream))
df_dt <- fit_c_ccf \%>\% calc_dt_CI(2 , new_data)
}

}
\author{
Priyanga Dilini Talagala & Puwasala Gamakumara
}
