% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_moments.R
\name{conditional_moments}
\alias{conditional_moments}
\title{Estimating conditional moments for time series}
\usage{
conditional_moments(
  data,
  x,
  z_numeric,
  z_factors = NULL,
  knots_mean = NULL,
  knots_variance = NULL
)
}
\arguments{
\item{data}{a tibble containing all the time series which are uniquely identified by the
corresponding Timestamp.}

\item{x}{variable for which the conditional moments are estimated. Should be given as bared/unquoted name}

\item{z_numeric}{numerical variable(s) use as predictors. Should be given as bared/unquoted names and use c()
for multiple variables}

\item{z_factors}{factor variable(s) use as predictors. Should be given as bared/unquoted names.
NULL for empty factors or use c() for multiple variables}

\item{knots_mean}{vector specifying the dimension of the basis in the smooth term fitting for
each predictor in the GAM for conditional mean of $x$. Each component of the vector should corresponds to each predictor specified in
"z_numeric". Default fitting a $3$ dimentional thin plate regression spline.}

\item{knots_variance}{a vector specifying the dimension of the basis in the smooth term fitting for
each predictor in the GAM for conditional variance of $x$. Each component of the vector should corresponds to each predictor specified in
"z_numeric". Default fitting a $3$ dimentional thin plate regression spline.}
}
\value{
an object of class "conditional_moments" with the following components
\item{data_conditional_moments}{The original tibble appended with the estimated
conditional means and variance of the time series given in $x$}
\item{x_mean_gam_fit}{fitted GAM model for conditional mean of $x$}
\item{x_var_gam_fit}{fitted GAM model for conditional variance of $x$}
}
\description{
This function estimates the means and variance of a time series conditional on a set of
other times series via additive models.
}
\details{
{ Suppose $x_t$ is a time series where its mean and the variance are functions of $z_t$.
i.e. $E(x_t|z_t) = m_x(z_t)$ and $Var(x_t|z_t) = v_x(z_t)$. Then $m_x(z_t)$ and $v_x(z_t)$ can be
estimated via generalised additive models (GAM). \code{conditional_moments} uses GAMs implemented
in \code{mgcv} package to estimate the conditional means and the variance of a time series given
a set of time series predictors.}
}
\seealso{
\code{\link[mgcv]{gam}}
}
\author{
Puwasala Gamakumara
}
