% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dt.R
\name{estimate_dt}
\alias{estimate_dt}
\title{Estimating time delay between two sensors in a river system}
\usage{
estimate_dt(object, new_data, k_min = NULL, k_max = NULL)
}
\arguments{
\item{object}{an object returned from \code{conditional_ccf} function}

\item{new_data}{data for which the time delay "dt" should be calculated}

\item{k_min}{minimum lag to be considered when calculating maximum cross-correlation.
default is $min(k)$ where k is the sequence of lags defined in \code{conditional_ccf}}

\item{k_max}{maximum lag to be considered when calculating maximum cross-correlation.
default is $max(k)$ where k is the sequence of lags defined in \code{conditional_ccf}}
}
\value{
an object of class "conditional_lag" with the following components
\item{data_dt}{The original tibble passed to "new_data" appended with the estimated time lag "dt"
and corresponding cross-correlation at maximum lag}
\item{conditional_ccf_object}{The object passed to the \code{estimate_dt} function}
\item{k_min}{minimum lag}
\item{k_max}{maximum lag}
}
\description{
This function estimates the time that takes water to flow from an upstream location to a downstream
location conditional on the observed water-quality variables from the upstream sensor. That time lag is
defined as the lag that gives maximum cross-correlation conditional on upstream water-quality variables.
}
