% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_lags.R
\name{map_lags}
\alias{map_lags}
\title{Computing upstream variables according to the estimated time lag}
\usage{
map_lags(data, up, down, lag)
}
\arguments{
\item{data}{a tibble containing all downstream variables, upstream variables, estimated time lag
and corresponding Timestamp}

\item{up}{bared/unquoted name(s) of upstream variable(s)}

\item{down}{bared/unquoted name(s) of downstream variable(s)}

\item{lag}{bared/unquoted name corresponds to the estimated time lag}
}
\value{
a tibble with Timestamp, downstream variables and upstream lagged variables corresponds to
each timestamp
}
\description{
This function will compute the lag of upstream variables according to the estimated time
lag returned from \code{estimate_dt} function.
}
\author{
Puwasala Gamakumara
}
