% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEON_PRIN_5min_cleaned}
\alias{NEON_PRIN_5min_cleaned}
\title{Anomaly removed data for water quality variables aggregated at 5-minute intervals
from Pringle Creek, Texas.}
\format{
A data frame with water-quality variables, level and temperature
data:
\describe{
\item{\code{Timestamp}}{Timestamp}
\item{\code{site}}{site position}
\item{\code{conductance}}{specific conductance}
\item{\code{dissolvedOxygen}}{dissolved oxygen}
\item{\code{pH}}{pH}
\item{\code{chlorophyll}}{chlorophyll}
\item{\code{turbidity}}{turbidity}
\item{\code{fDOM}}{fDOM}
\item{\code{level}}{elevation of surface water}
\item{\code{temperature}}{temperature in surface water}
}
}
\usage{
NEON_PRIN_5min_cleaned
}
\description{
NEON_PRIN_5min_cleaned consists anomaly removed data for water quality variables
from upstream and downstream sensors in Pringle Creek in Texas for the period
spanning from 2019-07-01 to 2019-12-31 aggregated at 5-minute intervals.
}
\keyword{datasets}
