% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEON_PRIN_level_temp_5min}
\alias{NEON_PRIN_level_temp_5min}
\title{Raw data for level and temperature variables from Pringle Creek, Texas.}
\format{
A data frame with the following columns:
\describe{
\item{\code{Timestamp}}{Timestamp}
\item{\code{site}}{site position}
\item{\code{level}}{elevation of surface water}
\item{\code{temperature}}{temperature in surface water}
}
}
\source{
This data are retrieved from NEON Aquatic Instrument Data.
For further details, see \url{https://data.neonscience.org/data-products/DP1.20016.001}
and \url{https://data.neonscience.org/data-products/DP1.20053.001}.
}
\usage{
NEON_PRIN_level_temp_5min
}
\description{
NEON_PRIN_level_temp_5min consists raw data for level and temperature variables
from upstream and downstream sensors in Pringle Creek in Texas for the period
spanning from 2019-07-01 to 2019-12-31 at
5-minute intervals.
}
\keyword{datasets}
