% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEON_PRIN_waq}
\alias{NEON_PRIN_waq}
\title{Raw data for water quality variables from Pringle Creek, Texas.}
\format{
A data frame with six water-quality variables with associated
anomaly flags:
\describe{
\item{\code{Timestamp}}{Timestamp}
\item{\code{horizontalPosition}}{site position}
\item{\code{conductance}}{specific conductance}
\item{\code{dissolvedOxygen}}{dissolved oxygen}
\item{\code{pH}}{pH}
\item{\code{turbidity}}{turbidity}
\item{\code{chlorophyll}}{chlorophyll}
\item{\code{fDOM}}{fDOM}
\item{\code{conductanceAnomalyFlag}}{anomaly flag for conductance}
\item{\code{dissolvedOxygenAnomalyFlag}}{anomaly flag for dissolved oxygen}
\item{\code{pHAnomalyFlag}}{anomaly flag for pH}
\item{\code{turbidityAnomalyFlag}}{anomaly flag for turbidity}
\item{\code{chlorophyllAnomalyFlag}}{anomaly flag for chlorophyll}
\item{\code{fDOMAnomalyFlag}}{anomaly flag for fDOM}
\item{\code{turbidity_wiperAnomalyFlag}}{wiper anomaly flag for turbidity}
}
}
\source{
This data are retrieved from NEON Aquatic Instrument Data.
For further details, see \url{https://data.neonscience.org/data-products/DP1.20288.001}
}
\usage{
NEON_PRIN_waq
}
\description{
NEON_PRIN_waq consists raw data for water quality variables from upstream
and downstream sensors in Pringle Creek in Texas for the period
spanning from 2019-07-01 to 2019-12-31 at
1-minute intervals.
}
\keyword{datasets}
