% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.conditional_ccf}
\alias{predict.conditional_ccf}
\title{Predict method for conditional cross-correlation fits}
\usage{
\method{predict}{conditional_ccf}(object, new_data, ...)
}
\arguments{
\item{object}{an object returned from \code{conditional_ccf} function}

\item{new_data}{a tibble for which the conditional cross-correlations need to be predicted.
This tibble should contain all the predictors passed to \code{conditional_ccf} along with the Timestamp.}

\item{...}{ignored}
}
\value{
returns a tibble with the predicted conditional cross-correlations
at lags $k=1,2,...$
}
\description{
Compute predictions for conditional cross-correlations at lags $k=1,2,...$ from the fitted models
in \code{conditional_ccf} function.
}
\author{
Puwasala Gamakumara
}
