% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{.data}
\alias{augment}
\alias{:=}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom:reexports]{augment}}}

  \item{rlang}{\code{\link[rlang:dot-data]{.data}}, \code{\link[rlang:dyn-dots]{:=}}}
}}

