% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.conditional_moments}
\alias{summary.conditional_moments}
\title{Summarising the models fitted for conditional moments}
\usage{
\method{summary}{conditional_moments}(object, type = c("mean", "variance"), ...)
}
\arguments{
\item{object}{an object returned from \code{conditional_moments} function}

\item{type}{type of the moment to return the model summary. This can take one of
"mean" or "variance"}

\item{...}{further arguments passed to or from other methods}
}
\value{
returns an object of class summary.gam
}
\description{
Summary methods for the class \code{conditional_moments}. Function gives
the summary of Generalised Additive Models fitted for conditional means and
conditional variances using \code{conditional_moments} function.
}
\seealso{
\code{\link[mgcv]{gam}}, \code{\link{summary}}
}
