% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.conditional_moments}
\alias{autoplot.conditional_moments}
\title{Visualising the models fitted for conditional moments}
\usage{
\method{autoplot}{conditional_moments}(object, type = c("mean", "variance"), ...)
}
\arguments{
\item{object}{an object returned from \code{conditional_moments} function}

\item{type}{the type of moment want to visualise. This can take one of
"mean" or "variance"}

\item{...}{ignored}
}
\value{
returns plots visualising fitted gam model for conditional
mean or variance vs each predictor.
}
\description{
A function for visualising the fitted generalised additive models for
conditional means and variances from \code{conditional_moments} function
}
\seealso{
\code{\link[mgcv]{plot.gam}}, \code{\link[visreg]{visreg}}
}
\author{
Puwasala Gamakumara
}
