% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_ccf.R
\name{conditional_ccf}
\alias{conditional_ccf}
\title{Computing conditional cross-correlations at given lags}
\usage{
conditional_ccf(
  data,
  x,
  y,
  z_numeric,
  z_factors,
  family = c("Gamma", "lognormal"),
  k = 1:9,
  knots_mean = NULL,
  knots_variance = NULL,
  df_correlation = NULL
)
}
\arguments{
\item{data}{a tibble containing all the time series which are uniquely identified by the
corresponding Timestamp}

\item{x}{bared/unquoted name of the variable to be considered as $x_t$}

\item{y}{bared/unquoted name of the variable to be considered as $y_t$}

\item{z_numeric}{numerical variable(s) use as predictors. Should be given as bared/unquoted names
and use c() for multiple variables}

\item{z_factors}{factor variable(s) use as predictors. Should be given as bared/unquoted names.
NULL for empty factors or use c() for multiple variables}

\item{family}{the family to be used in conditional variance model. Currently
this can take either "Gamma" or "lognormal".}

\item{k}{a vector of lag values at which the cross-correlation needs to be computed. Default is
$1:9$}

\item{knots_mean}{a named list of vectors specifiying the dimension of the basis of the smooth term fitting for
each predictor in the conditional mean models for $x_t$ and $y_t$ (see \code{conditional_moments}). The vectors should be named as
$x$ and $y$. The components of each vector should correspond to each predictor specified in
"z_numeric". By default function fits a 3 dimentional thin plate regression spline for each predictor.}

\item{knots_variance}{a named list of vectors specifiying the dimension of the basis of the smooth term fitting for
each predictor in the conditional variance models for $x_t$ and $y_t$ (see \code{conditional_moments}). The vectors should be named as
$x$ and $y$. The components of each vector should correspond to each predictor specified in
"z_numeric". By default the function fits a 3 dimentional thin plate regression spline for each predictor.}

\item{df_correlation}{a vector specifying the degrees of freedom to be considered for each numerical
predictor when fitting additive models for conditional cross-correlations. Each component of the
vector should corresponds to each predictor specified in "z_numeric". By default the function
will fit a natural cubic spline with $3$ degrees of freedom. see \code{\link[splines]{ns}}
consider $3$ degrees for each predictor.}
}
\value{
an object of class "conditional_ccf" with the following components
\item{data_ccf}{The original tibble passed to the function appended with the estimated
conditional cross correlation at lags $k = 1,2,...$}
\item{data_visualise}{a list containing the fitted models for conditional means, variance and
cross-correlation at each lag. These data can be used for visualisation and diagnostics of each fitted model}
\item{formula_gam}{formula passed to the \code{\link[stats]{glm}} to fit additive models for
conditional cross-correlation}
}
\description{
This function computes cross correlation between $x_t$ and $y_{t+k}$ at $k = 1,2,...$
conditional on a set of time series $z_t$
}
\details{
{ Suppose $x_t$ and $y_t$ are conditionally normalised with respect
to $z_t$ using \code{conditional_moments}. Then we can estimate the conditional
cross-correlation between $x_t$ and $y_t$ at lag $k$, i.e. $r_k = E(x_ty_{t+k}|z_t)$
via generalised additive models (GAM). \code{conditional_ccf} uses natural splines implemented
in \code{splines} package to estimate the conditional cross-correlations between two
time series given a set of time series predictors. Users need not to
normalise $x_t$ and $y_t$. The function \code{conditional_ccf} itself will
normalise $x_t$ and $y_t$ using \code{conditional_moments}}
}
\seealso{
\code{\link[stats]{glm}}, \code{\link[splines]{ns}}
}
\author{
Puwasala Gamakumara
}
