% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_mean.R
\name{conditional_mean}
\alias{conditional_mean}
\title{Estimating conditional mean of a time series}
\usage{
conditional_mean(data, formula)
}
\arguments{
\item{data}{a tibble containing all the time series
which are uniquely identified by the corresponding
Timestamp.}

\item{formula}{A GAM formula. See \code{\link[mgcv]{formula.gam}}.
The details of model specification are given under ‘Details’.}
}
\value{
The function returns an object of class
"gam" as described in \code{\link[mgcv]{gamObject}}.
}
\description{
This function estimates the means  of a time
series conditional on a set of  other times series
via additive models.
}
\details{
{ Suppose $x_t$ is a time series where its
mean is a function of $z_t$. i.e. $E(x_t|z_t) = m_x(z_t)$.
Then $m_x(z_t)$ can be estimated via generalised
additive models (GAM). This function uses
GAMs implemented in \code{mgcv} package to estimate
the conditional means of a time series given a set of
time series predictors.}
}
\examples{
data <- NEON_PRIN_5min_cleaned |>
  dplyr::filter(site == "upstream") |>
  dplyr::select(Timestamp, turbidity, level, conductance, temperature)

fit_mean <- data |>
  conditional_mean(turbidity ~ s(level, k = 8) +
    s(conductance, k = 8) + s(temperature, k = 8))
}
\seealso{
\code{\link[mgcv]{gam}}
}
