% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduits.R
\docType{package}
\name{conduits}
\alias{conduits}
\title{conduits: CONDitional User Interface for Time Series normalisation}
\description{
Methods and tools for conditional normalisation of time series using additive models.
This includes functions to estimate conditional autocorrelation functions and conditional cross-correlation functions.
Further these functions can be used to estimate river flow time between two sensor locations in a river system.
}
\author{
Puwasala Gamakumara
}
