% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a series using conditional moments}
\usage{
normalize(data, y, fit_mean, fit_var)
}
\arguments{
\item{data}{a tsibble containing all the time series
which are uniquely identified by the corresponding
Timestamp.}

\item{y}{The variable name}

\item{fit_mean}{Model object of class "conditional_moment" returned from
\code{\link[conduits]{conditional_mean}} with information to append to observations.}

\item{fit_var}{Model object of class "conditional_moment" returned from
\code{\link[conduits]{conditional_var}} with information to append to observations.}
}
\value{
A vector of conditional normliased series
}
\description{
This function produces  a normalized series using conditional moments.
}
\examples{
data <- NEON_PRIN_5min_cleaned \%>\%
  dplyr::filter(site == "upstream") \%>\%
  dplyr::select(Timestamp, turbidity, level, conductance, temperature) \%>\%
  tsibble::as_tsibble(index = Timestamp)

fit_mean <- data \%>\%
  conditional_mean(turbidity ~ s(level, k = 8) +
    s(conductance, k = 8) + s(temperature, k = 8))

fit_var <- data \%>\%
  conditional_var(
    turbidity ~ s(level, k = 7) + s(conductance, k = 7) + s(temperature, k = 7),
    family = "Gamma",
    fit_mean
  )

new_ts <- data \%>\%
 dplyr::mutate(
 ystar = conduits::normalize(., turbidity, fit_mean, fit_var))

}
